Policies post

+++ app/Http/Controllers/PostController.php
@@ -39,20 +39,30 @@ class PostController extends Controller
         return redirect()->route("post.index")->with("success", "Publication ajoutee");
     }
 
-    public function edit(Post $post)
+    public function edit(Post $post, Request $request)
     {
+        if($request->user()->cannot("update", $post)) {
+            abort(403);
+        }
+
         return view("post.edit", compact("post"));
     }
 
     public function update(StorePostFormRequest $request, Post $post)
     {
+         if($request->user()->cannot("update", $post)) {
+            abort(403);
+        }
         $data = $request->validated();
         $post->update($data);
         return redirect()->route("post.show", $post)->with("success", "Publication Modifiee");
     }
 
-    public function destroy(Post $post)
+    public function destroy(Post $post, Request $request)
     {
+         if($request->user()->cannot("delete", $post)) {
+            abort(403);
+        }
         $post->delete();
         return redirect()->route("post.index")->with("success", "Publication Supprimee");
     }
